+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added
 Changed
   * AZ-740: System assigned identity configured by default
   * AZ-740: Zone redundancy enabled by default
+  * AZ-739: Remove validation constraint on `mongo_server_version` variable and set default value to `4.2`
 
 # v6.0.0 - 2022-05-11